+++ garmin_pipeline_parallel.py
@@ -1073,7 +1073,7 @@ def pipeline(study_id: str, workers: int = 4, args: list = None):
 if __name__ == "__main__":
     sys_args = sys.argv
 
-    workers = 4
+    workers = 1
 
     parser = argparse.ArgumentParser(description="Process garmin data files")
     parser.add_argument(
